+++ src/main/java/dev/orne/beans/LongIdentity.java
@@ -51,6 +51,17 @@ extends AbstractSimpleIdentity<Long> {
     }
 
     /**
+     * Copy constructor.
+     * 
+     * @param copy The instance to copy
+     */
+    public LongIdentity(
+            @Nonnull
+            final LongIdentity copy) {
+        super(copy);
+    }
+
+    /**
      * Resolves the specified identity token to a valid {@code StringIdentity}
      * 
      * @param token The identity token
@@ -69,7 +80,7 @@ extends AbstractSimpleIdentity<Long> {
                 IdentityTokenFormatter.DEFAULT_PREFIX,
                 token);
         if (body == null) {
-            return new LongIdentity(null);
+            return new LongIdentity((Long) null);
         } else {
             try {
                 return new LongIdentity(Long.parseLong(body));