+++ src/main/java/dev/orne/beans/StringIdentity.java
@@ -51,6 +51,17 @@ extends AbstractSimpleIdentity<String> {
     }
 
     /**
+     * Copy constructor.
+     * 
+     * @param copy The instance to copy
+     */
+    public StringIdentity(
+            @Nonnull
+            final StringIdentity copy) {
+        super(copy);
+    }
+
+    /**
      * Resolves the specified identity token to a valid {@code StringIdentity}
      * 
      * @param token The identity token