+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+# :package: 0.2.0
+
+01. Improved class `AbstractIdentity`
+    01. :gift: Method `toString` returns identity token
+    01. :gift: Generated identity token is cached
+01. Improved class `AbstractSimpleIdentity`
+    01. :gift: Added copy constructor
+    01. :gift: Implemented `hashCode` and `equals`
+01. Improved class `StringIdentity`
+    01. :gift: Added copy constructor
+01. Improved class `LongIdentity`
+    01. :gift: Added copy constructor
+01. Improved class `BigIntegerIdentity`
+    01. :gift: Added copy constructor
+
 # :package: 0.1.0
 
 01. :gift: Added exception `dev.orne.beans.UnrecognizedIdentityTokenException`